Harmonize webflux sample test
