+++ spring-boot-samples/spring-boot-sample-webflux/src/test/java/sample/webflux/SampleWebFluxApplicationTests.java
@@ -19,9 +19,12 @@ package sample.webflux;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.test.context.SpringBootTest.WebEnvironment;
+import org.springframework.http.MediaType;
 import org.springframework.test.context.junit4.SpringRunner;
+import org.springframework.test.web.reactive.server.WebTestClient;
 
 /**
  * Basic integration tests for WebFlux application.
@@ -29,12 +32,16 @@ import org.springframework.test.context.junit4.SpringRunner;
  * @author Brian Clozel
  */
 @RunWith(SpringRunner.class)
-@SpringBootTest(webEnvironment = WebEnvironment.NONE)
+@SpringBootTest(webEnvironment = WebEnvironment.RANDOM_PORT)
 public class SampleWebFluxApplicationTests {
 
-	@Test
-	public void testSomething() throws Exception {
+	@Autowired
+	private WebTestClient webClient;
 
+	@Test
+	public void testWelcome() throws Exception {
+		this.webClient.get().uri("/").accept(MediaType.TEXT_PLAIN).exchange()
+				.expectBody(String.class).value().isEqualTo("Hello World");
 	}
 
 }
